refactored navBar
